# Optimize account lookup by `auth-addr`